move equal files to shared source
