+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/Server/ITypedGetters.cs
