+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/Server/SmiXetterTypeCode.cs
