+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/Server/SqlMethodAttribute.cs
@@ -6,7 +6,7 @@ using System;
 
 namespace Microsoft.Data.SqlClient.Server
 {
-    /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlMethodAttribute.xml' path='docs/members[@name="SqlMethodAttribute"]/SqlMethodAttribute/*' />
+    /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlMethodAttribute.xml' path='docs/members[@name="SqlMethodAttribute"]/SqlMethodAttribute/*' />
     [AttributeUsage(AttributeTargets.Method, AllowMultiple = false, Inherited = false), Serializable]
     public sealed class SqlMethodAttribute : SqlFunctionAttribute
     {
@@ -14,7 +14,7 @@ namespace Microsoft.Data.SqlClient.Server
         private bool _isMutator;
         private bool _shouldInvokeIfReceiverIsNull;
 
-        /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlMethodAttribute.xml' path='docs/members[@name="SqlMethodAttribute"]/ctor/*' />
+        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlMethodAttribute.xml' path='docs/members[@name="SqlMethodAttribute"]/ctor/*' />
         public SqlMethodAttribute()
         {
             // default values
@@ -23,21 +23,21 @@ namespace Microsoft.Data.SqlClient.Server
             _shouldInvokeIfReceiverIsNull = false;
         }
 
-        /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlMethodAttribute.xml' path='docs/members[@name="SqlMethodAttribute"]/OnNullCall/*' />
+        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlMethodAttribute.xml' path='docs/members[@name="SqlMethodAttribute"]/OnNullCall/*' />
         public bool OnNullCall
         {
             get => _isCalledOnNullInputs;
             set => _isCalledOnNullInputs = value;
         }
 
-        /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlMethodAttribute.xml' path='docs/members[@name="SqlMethodAttribute"]/IsMutator/*' />
+        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlMethodAttribute.xml' path='docs/members[@name="SqlMethodAttribute"]/IsMutator/*' />
         public bool IsMutator
         {
             get => _isMutator;
             set => _isMutator = value;
         }
 
-        /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlMethodAttribute.xml' path='docs/members[@name="SqlMethodAttribute"]/InvokeIfReceiverIsNull/*' />
+        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlMethodAttribute.xml' path='docs/members[@name="SqlMethodAttribute"]/InvokeIfReceiverIsNull/*' />
         public bool InvokeIfReceiverIsNull
         {
             get => _shouldInvokeIfReceiverIsNull;
