+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/Server/SqlUserDefinedTypeAttribute.cs
@@ -7,18 +7,18 @@ using Microsoft.Data.Common;
 
 namespace Microsoft.Data.SqlClient.Server
 {
-    /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/Format.xml' path='docs/members[@name="Format"]/Format/*' />
+    /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/Format.xml' path='docs/members[@name="Format"]/Format/*' />
     public enum Format
     {
-        /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/Format.xml' path='docs/members[@name="Format"]/Unknown/*' />
+        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/Format.xml' path='docs/members[@name="Format"]/Unknown/*' />
         Unknown = 0,
-        /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/Format.xml' path='docs/members[@name="Format"]/Native/*' />
+        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/Format.xml' path='docs/members[@name="Format"]/Native/*' />
         Native = 1,
-        /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/Format.xml' path='docs/members[@name="Format"]/UserDefined/*' />
+        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/Format.xml' path='docs/members[@name="Format"]/UserDefined/*' />
         UserDefined = 2,
     }
 
-    /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedTypeAttribute.xml' path='docs/members[@name="SqlUserDefinedTypeAttribute"]/SqlUserDefinedTypeAttribute/*' />
+    /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedTypeAttribute.xml' path='docs/members[@name="SqlUserDefinedTypeAttribute"]/SqlUserDefinedTypeAttribute/*' />
     // This custom attribute indicates that the given type is
     // a SqlServer udt. The properties on the attribute reflect the
     // physical attributes that will be used when the type is registered
@@ -36,7 +36,7 @@ namespace Microsoft.Data.SqlClient.Server
         internal const int YukonMaxByteSizeValue = 8000;
         private string _validationMethodName = null;
 
-        /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedTypeAttribute.xml' path='docs/members[@name="SqlUserDefinedTypeAttribute"]/ctor/*' />
+        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedTypeAttribute.xml' path='docs/members[@name="SqlUserDefinedTypeAttribute"]/ctor/*' />
         // A required attribute on all udts, used to indicate that the
         // given type is a udt, and its storage format.
         public SqlUserDefinedTypeAttribute(Format format)
@@ -54,7 +54,7 @@ namespace Microsoft.Data.SqlClient.Server
             }
         }
 
-        /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedTypeAttribute.xml' path='docs/members[@name="SqlUserDefinedTypeAttribute"]/MaxByteSize/*' />
+        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedTypeAttribute.xml' path='docs/members[@name="SqlUserDefinedTypeAttribute"]/MaxByteSize/*' />
         // The maximum size of this instance, in bytes. Does not have to be
         // specified for Native serialization. The maximum value
         // for this property is specified by MaxByteSizeValue.
@@ -74,7 +74,7 @@ namespace Microsoft.Data.SqlClient.Server
             }
         }
 
-        /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedTypeAttribute.xml' path='docs/members[@name="SqlUserDefinedTypeAttribute"]/IsFixedLength/*' />
+        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedTypeAttribute.xml' path='docs/members[@name="SqlUserDefinedTypeAttribute"]/IsFixedLength/*' />
         // Are all instances of this udt the same size on disk?
         public bool IsFixedLength
         {
@@ -88,7 +88,7 @@ namespace Microsoft.Data.SqlClient.Server
             }
         }
 
-        /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedTypeAttribute.xml' path='docs/members[@name="SqlUserDefinedTypeAttribute"]/IsByteOrdered/*' />
+        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedTypeAttribute.xml' path='docs/members[@name="SqlUserDefinedTypeAttribute"]/IsByteOrdered/*' />
         // Is this type byte ordered, i.e. is the on disk representation
         // consistent with the ordering semantics for this type?
         // If true, the binary representation of the type will be used
@@ -106,11 +106,11 @@ namespace Microsoft.Data.SqlClient.Server
             }
         }
 
-        /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedTypeAttribute.xml' path='docs/members[@name="SqlUserDefinedTypeAttribute"]/Format/*' />
+        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedTypeAttribute.xml' path='docs/members[@name="SqlUserDefinedTypeAttribute"]/Format/*' />
         // The on-disk format for this type.
         public Format Format => _format;
 
-        /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedTypeAttribute.xml' path='docs/members[@name="SqlUserDefinedTypeAttribute"]/ValidationMethodName/*' />
+        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedTypeAttribute.xml' path='docs/members[@name="SqlUserDefinedTypeAttribute"]/ValidationMethodName/*' />
         // An Optional method used to validate this UDT
         public string ValidationMethodName
         {
@@ -124,7 +124,7 @@ namespace Microsoft.Data.SqlClient.Server
             }
         }
 
-        /// <include file='../../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedTypeAttribute.xml' path='docs/members[@name="SqlUserDefinedTypeAttribute"]/Name/*' />
+        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient.Server/SqlUserDefinedTypeAttribute.xml' path='docs/members[@name="SqlUserDefinedTypeAttribute"]/Name/*' />
         public string Name
         {
             get
